EMSUSD-3695 add comment and move util
- Added comments why teh setVisibility uses the `NoUsdUndoBlockGuard`
- Move the guard to a better header.
