+++ lib/mayaUsd/editForward/MayaUsdEditForwardHost.cpp
@@ -22,7 +22,7 @@
 
 #include <usdUfe/base/debugCodes.h>
 #include <usdUfe/ufe/UsdUndoableCommand.h>
-#include <usdUfe/ufe/trf/Utils.h>
+#include <usdUfe/ufe/Utils.h>
 #include <usdUfe/undo/UsdUndoBlock.h>
 #include <usdUfe/undo/UsdUndoManager.h>
