+++ lib/usdUfe/ufe/trf/Utils.h
@@ -54,35 +54,6 @@ void rotatePivotTranslateOp(
     double                 y,
     double                 z);
 
-//! Guard to set and reset a flag indicating that we are in a command
-//! that does not use a UsdUndoBlock but still wants edit-forwarding.
-class USDUFE_PUBLIC NoUsdUndoBlockGuard
-{
-public:
-    NoUsdUndoBlockGuard(bool noUsdUndoBlock)
-        : _noUsdUndoBlock(noUsdUndoBlock)
-    {
-        if (_noUsdUndoBlock)
-            _getGuardedFlag()++;
-    }
-    ~NoUsdUndoBlockGuard()
-    {
-        if (_noUsdUndoBlock)
-            _getGuardedFlag()--;
-    }
-
-    static bool wantsForwarding() { return _getGuardedFlag() > 0; }
-
-private:
-    static int& _getGuardedFlag()
-    {
-        static int flag = 0;
-        return flag;
-    }
-
-    bool _noUsdUndoBlock;
-};
-
 } // namespace USDUFE_NS_DEF
 
 #endif // USDUFE_UFE_TRF_UTILS_H
